requirements.txt
# Defender Capital Management — Python Requirements
# Install with: pip install -r requirements.txt
# Or use setup.sh / setup.bat for a full virtual environment setup

openpyxl>=3.1.0
pandas>=2.0.0
numpy>=1.26.0
yfinance>=0.2.36
scipy>=1.11.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0
kaleido>=0.2.1
streamlit>=1.41.0
requests>=2.31.0
python-dateutil>=2.8.2
jinja2>=3.1.0
pytest>=7.4.0
psycopg2-binary>=2.9.9
bcrypt>=4.1.0
